backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.3
email-validator==2.1.0
pytz==2024.1
gunicorn
load_dotenv